requirements.txt
fastapi==0.103.0
numpy==1.25.1
opencv_python==4.8.0.76
Pillow==10.0.0
Requests==2.31.0
torch==2.0.1
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.66.1
uvicorn==0.23.2
